move sight list widgets to widgets folder
