+++ lib/ui/screen/sight_list_screen/sight_list_screen.dart
@@ -6,9 +6,9 @@ import 'package:places/constants/app_strings.dart';
 import 'package:places/data/providers/sights_api.dart';
 import 'package:places/styles/custom_icons.dart';
 import 'package:places/ui/screen/add_sight_screen/add_sight_widget.dart';
-import 'package:places/ui/screen/sight_list_screen/sight_list_landscape_orientation.dart';
-import 'package:places/ui/screen/sight_list_screen/sight_list_portrate_orientation.dart';
-import 'package:places/ui/screen/sight_list_screen/sight_list_search_bar_delegate.dart';
+import 'package:places/ui/screen/sight_list_screen/widgets/sight_list_landscape_orientation.dart';
+import 'package:places/ui/screen/sight_list_screen/widgets/sight_list_portrate_orientation.dart';
+import 'package:places/ui/screen/sight_list_screen/widgets/sight_list_search_bar_delegate.dart';
 import 'package:places/ui/screen/sight_search_screen/sight_search_screen.dart';
 import 'package:places/ui/widgets/buttons/rounded_gradient_button.dart';
 import 'package:places/ui/widgets/error_pages/network_error_page.dart';
