+++ lib/ui/screen/sight_list_screen/widgets/sight_list_portrate_orientation.dart
@@ -1,19 +1,21 @@
 import 'package:flutter/material.dart';
 
-import '../../../data/model/sights/sight.dart';
-import '../../widgets/sight_cards/sight_card.dart';
+import '../../../../data/model/sights/sight.dart';
+import '../../../widgets/sight_cards/sight_card.dart';
 
-class SightListLandscapeOrientation extends StatelessWidget {
-  const SightListLandscapeOrientation(this.sights);
+class SightListPortrateOrientation extends StatelessWidget {
+  const SightListPortrateOrientation(this.sights);
 
   final List<Sight> sights;
 
   @override
   Widget build(BuildContext context) {
-    return SliverGrid(
+    return SliverList(
       delegate: SliverChildBuilderDelegate(
         (context, index) {
           double bottomPadding = 0;
+          bool isLastElement = index == sights.length - 1;
+          if (isLastElement) bottomPadding = 80;
 
           return Padding(
             padding: EdgeInsets.only(
@@ -28,10 +30,6 @@ class SightListLandscapeOrientation extends StatelessWidget {
           );
         },
         childCount: sights.length,
-      ),
-      gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-        crossAxisCount: 2,
-        childAspectRatio: 1.6,
       ),
     );
   }
